Rename two of the variables
Rename V to nextTimestepMatrix, and nextTimestep to
nextTimestepVector.

Due to the nature of the formulation of the scheme, it has a
particular way of ordering the previous timestep points and the
next timestep points; they are expressed as vectors rather than
matrices.

Therefore, in order to express the solution in terms of matrices
(which is the most natural way of describing a 2D problem) we
must rearrange the vector output of the scheme into a matrix of
the appropriate dimensions.

With that in mind, the names of two of the variables, V and
nextTimestep, have been changed to try and make their roles
in the scheme more obvious.