Edit ordering and application of BCs
Move the application of the y BCs before the x BCs, and only apply
the y BCs to the inner points along the boundary and letting the
x BCs determine the corner points.

For Dirichlet BCs this isn't necessary, but the ordering of the BCs
as well as which BCs are used for finding the solutions of the
corner points becomes important when considering Neumann BCs, so this
is a preparation step for the later generalisations to include
Neumann BCs as well as Dirichlet BCs.

Could have equally chosen to do it the other way around, using the x
BCs for the inner points and using the y BCs to get the corner points,
just arbitrarily chose to do it this way.