Allow choice to use Neumann BCs
Use one-sided scheme to solve for boundary points which have a
Neumann BC associated with them.

For now, only allow both boundaries for x or y to be one BC
condition or the other, they cannot be mixed (ie, if the LHS
BC for x is Dirichlet then the RHS cannot be Neumann, it
must also be Dirichlet).