Change value of dt
An error I made when calculating a condition for von Neumann
stability has shown that a smaller value of dt is required for
stability of the scheme.

In particular, the old (incorrect) requirement for stability I
had was 1 >= dt * (1/(dx^2) + 1/(dy^2)), but the actual requirement
for stability is 1 >= 2 * dt * (1/(dx^2) + 1/(dy^2)), and it can
easily be seen from this that the old value of dt = 0.025 is
too large to satisfy the correct stability condition.

Therefore, a new value of dt = 0.005 has been arbitrarily chosen
as a satisfactory value to ensure stability of the FTCS scheme.