Add FiniteTimePDESolver class
Move the non-core functionalities of PDESolver into a subclass,
FiniteTimePDESolver.

The purposes of this are largely tied to the different uses of
a numerical PDE solver; I imagine that it's split into 2 main
groups:
1. obtaining data for a purely mathematical interest in the PDE's
solution
2. obtaining data to be fed into another program

The main distinction between these two groups are that group 1
likely are interested in a finite time interval, whilst group 2
will more likely want real time data for an indefinite period
of time.

Moreover, the overlap in extra functionalities between these two
groups are minimal, with only the mathematical routines needed for
numerically solving the PDE really being the common ground.

Therefore, having PDESolver as a base class and then having two
subclasses of it, one for group 1 and one for group 2, seems
like a reasonable way to achieve the functionalities required
by most users interested, while also maintaining a good level
of separation to not give more functionalities to one group
than is needed or wanted.

FiniteTimePDESolver is for group 1, and hopefully there will be
a subclass of PDESolver for group 2 soon.