Add function to IndefiniteTimePDESolver class
Another feature that could be useful for users who want to use the
PDE solution data as the input of some other program is to be able
to have realtime interactions with the system being modelled.

Therefore, the set_next_timestep_matrix_entry function has been
added to IndefiniteTimePDESolver as an attempt to enable
interactions, with the idea being that the new value of the entry
you want to change is a function of the old value in some way.

Interactions would then be performed by catching user input of
some form and then calling set_next_timestep_matrix_entry at some
point when handling this input.